Extend Insight FE model with capture group
