+++ client/web/src/enterprise/insights/components/insights-view-grid/components/smart-insight/SmartInsight.tsx
@@ -3,8 +3,8 @@ import React, { forwardRef, ReactElement, Ref } from 'react'
 import { ViewContexts } from '@sourcegraph/shared/src/api/extension/extensionHostApi'
 import { TelemetryProps } from '@sourcegraph/shared/src/telemetry/telemetryService'
 
-import { Insight, isBackendInsight, isCaptureGroupInsight, isSearchBasedInsight } from '../../../../core/types'
-import { BackendInsight } from '../backend-insight/BackendInsight'
+import { Insight, isBackendInsight } from '../../../../core/types'
+import { BackendInsightView } from '../backend-insight/BackendInsight'
 import { BuiltInInsight } from '../built-in-insight/BuiltInInsight'
 
 export interface SmartInsightProps<D extends keyof ViewContexts>
@@ -24,9 +24,9 @@ export interface SmartInsightProps<D extends keyof ViewContexts>
 export const SmartInsight = forwardRef<HTMLElement, SmartInsightProps<keyof ViewContexts>>((props, reference) => {
     const { insight, resizing = false, telemetryService, where, context, ...otherProps } = props
 
-    if (isSearchBasedInsight(insight) && isBackendInsight(insight)) {
+    if (isBackendInsight(insight)) {
         return (
-            <BackendInsight
+            <BackendInsightView
                 insight={insight}
                 resizing={resizing}
                 telemetryService={telemetryService}
@@ -34,11 +34,6 @@ export const SmartInsight = forwardRef<HTMLElement, SmartInsightProps<keyof View
                 innerRef={reference}
             />
         )
-    }
-
-    if (isCaptureGroupInsight(insight)) {
-        // TODO: Will be implemented in a separate PR about connecting capture group UI.
-        return null
     }
 
     // Search based extension and lang stats insight are handled by built-in fetchers
