+++ client/web/src/enterprise/insights/core/backend/gql-api/code-insights-gql-backend.ts
@@ -1,5 +1,5 @@
 import { ApolloCache, ApolloClient, gql } from '@apollo/client'
-import { from, Observable, of, throwError } from 'rxjs'
+import { from, Observable, of } from 'rxjs'
 import { map, mapTo, switchMap } from 'rxjs/operators'
 import { LineChartContent, PieChartContent } from 'sourcegraph'
 import {
@@ -9,24 +9,18 @@ import {
     DeleteDashboardResult,
     GetDashboardInsightsResult,
     GetInsightsResult,
-    GetInsightViewResult,
     InsightsDashboardsResult,
     InsightSubjectsResult,
-    InsightViewFiltersInput,
     RemoveInsightViewFromDashboardResult,
     UpdateDashboardResult,
     UpdateInsightsDashboardInput,
-    UpdateLangStatsInsightResult,
-    UpdateLangStatsInsightVariables,
-    UpdateLineChartSearchInsightResult,
 } from 'src/graphql-operations'
 
 import { ViewContexts } from '@sourcegraph/shared/src/api/extension/extensionHostApi'
 import { fromObservableQuery } from '@sourcegraph/shared/src/graphql/apollo'
 
-import { Insight, InsightDashboard, InsightsDashboardScope, InsightsDashboardType, InsightType } from '../../types'
+import { BackendInsight, Insight, InsightDashboard, InsightsDashboardScope, InsightsDashboardType } from '../../types'
 import { ALL_INSIGHTS_DASHBOARD_ID } from '../../types/dashboard/virtual-dashboard'
-import { SearchBackendBasedInsight } from '../../types/insight/search-insight'
 import { SupportedInsightSubject } from '../../types/subjects'
 import { CodeInsightsBackend } from '../code-insights-backend'
 import {
@@ -48,23 +42,18 @@ import { getLangStatsInsightContent } from '../core/api/get-lang-stats-insight-c
 import { getRepositorySuggestions } from '../core/api/get-repository-suggestions'
 import { getResolvedSearchRepositories } from '../core/api/get-resolved-search-repositories'
 import { getSearchInsightContent } from '../core/api/get-search-insight-content/get-search-insight-content'
-import { createLineChartContentFromIndexedSeries } from '../utils/create-line-chart-content'
-import { InsightInProcessError } from '../utils/errors'
 import { parseDashboardScope } from '../utils/parse-dashboard-scope'
 
+import { createInsightView } from './deserialization/create-insight-view'
 import { GET_DASHBOARD_INSIGHTS_GQL } from './gql/GetDashboardInsights'
 import { GET_INSIGHTS_GQL } from './gql/GetInsights'
 import { GET_INSIGHTS_DASHBOARDS_GQL } from './gql/GetInsightsDashboards'
 import { GET_INSIGHTS_SUBJECTS_GQL } from './gql/GetInsightSubjects'
-import { GET_INSIGHT_VIEW_GQL } from './gql/GetInsightView'
-import { UPDATE_LANG_STATS_INSIGHT_GQL } from './gql/UpdateLangStatsInsight'
-import { UPDATE_LINE_CHART_SEARCH_INSIGHT_GQL } from './gql/UpdateLineChartSearchInsight'
-import { createInsight } from './methods/create-insight'
+import { createInsight } from './methods/create-insight/create-insight'
+import { getBackendInsightData } from './methods/get-backend-insight-data/get-backend-insight-data'
 import { getCaptureGroupInsightsPreview } from './methods/get-capture-group-insight-preivew'
-import { getCaptureGroupInsightUpdateInput } from './serialization/capture-insight-to-gql-input'
-import { getSearchInsightUpdateInput } from './serialization/search-insight-to-gql-input'
+import { updateInsight } from './methods/update-insight/update-insight'
 import { createDashboardGrants } from './utils/get-dashboard-grants'
-import { getInsightView } from './utils/insight-transformers'
 
 export class CodeInsightsGqlBackend implements CodeInsightsBackend {
     constructor(private apolloClient: ApolloClient<object>) {}
@@ -77,10 +66,8 @@ export class CodeInsightsGqlBackend implements CodeInsightsBackend {
         // we need to use here insightViews query to fetch all available insights
         if (dashboardId === ALL_INSIGHTS_DASHBOARD_ID) {
             return fromObservableQuery(
-                this.apolloClient.watchQuery<GetInsightsResult>({
-                    query: GET_INSIGHTS_GQL,
-                })
-            ).pipe(map(({ data }) => data.insightViews.nodes.map(getInsightView).filter(Boolean) as Insight[]))
+                this.apolloClient.watchQuery<GetInsightsResult>({ query: GET_INSIGHTS_GQL })
+            ).pipe(map(({ data }) => data.insightViews.nodes.map(createInsightView)))
         }
 
         // Get all insights from the user-created dashboard
@@ -89,7 +76,10 @@ export class CodeInsightsGqlBackend implements CodeInsightsBackend {
                 query: GET_DASHBOARD_INSIGHTS_GQL,
                 variables: { id: dashboardId },
             })
-        ).pipe(map(({ data }) => data.insightsDashboards.nodes[0].views?.nodes.map(getInsightView) ?? []))
+        ).pipe(
+            map(({ data }) => data.insightsDashboards.nodes[0]),
+            map(dashboard => dashboard.views?.nodes.map(createInsightView) ?? [])
+        )
     }
 
     public getInsightById = (id: string): Observable<Insight | null> =>
@@ -106,7 +96,7 @@ export class CodeInsightsGqlBackend implements CodeInsightsBackend {
                     return null
                 }
 
-                return getInsightView(insightData) || null
+                return createInsightView(insightData) || null
             })
         )
 
@@ -128,46 +118,8 @@ export class CodeInsightsGqlBackend implements CodeInsightsBackend {
             )
         )
 
-    // TODO: Rethink all of this method. Currently `createViewContent` expects a different format of
-    // the `Insight` type than we use elsewhere. This is a temporary solution to make the code
-    // fit with both of those shapes.
-    public getBackendInsightData = (insight: SearchBackendBasedInsight): Observable<BackendInsightData> => {
-        const filters: InsightViewFiltersInput = {
-            includeRepoRegex: insight.filters?.includeRepoRegexp,
-            excludeRepoRegex: insight.filters?.excludeRepoRegexp,
-        }
-
-        return from(
-            // TODO: Use watchQuery instead of query when setting migration api is deprecated
-            this.apolloClient.query<GetInsightViewResult>({
-                query: GET_INSIGHT_VIEW_GQL,
-                variables: { id: insight.id, filters },
-            })
-        ).pipe(
-            // Note: this insight is guaranteed to exist since this function
-            // is only called from within a loop of insight ids
-            map(({ data }) => data.insightViews.nodes[0]),
-            switchMap(data => {
-                if (!data) {
-                    return throwError(new InsightInProcessError())
-                }
-
-                return of(data)
-            }),
-            map(data => ({
-                id: insight.id,
-                view: {
-                    title: insight.title ?? insight.title,
-                    // TODO: is this still used anywhere?
-                    subtitle: '',
-                    content: [createLineChartContentFromIndexedSeries(data.dataSeries, insight.series)],
-                    isFetchingHistoricalData: data.dataSeries.some(
-                        ({ status: { pendingJobs, backfillQueuedAt } }) => pendingJobs > 0 || backfillQueuedAt === null
-                    ),
-                },
-            }))
-        )
-    }
+    public getBackendInsightData = (insight: BackendInsight): Observable<BackendInsightData> =>
+        getBackendInsightData(this.apolloClient, insight)
 
     public getBuiltInInsightData = getBuiltInInsight
 
@@ -177,53 +129,7 @@ export class CodeInsightsGqlBackend implements CodeInsightsBackend {
 
     public createInsight = (input: InsightCreateInput): Observable<unknown> => createInsight(this.apolloClient, input)
 
-    public updateInsight = (input: InsightUpdateInput): Observable<unknown> => {
-        const insight = input.newInsight
-        const oldInsight = input.oldInsight
-
-        switch (insight.viewType) {
-            case InsightType.SearchBased: {
-                const input = getSearchInsightUpdateInput(insight)
-
-                return from(
-                    this.apolloClient.mutate<UpdateLineChartSearchInsightResult>({
-                        mutation: UPDATE_LINE_CHART_SEARCH_INSIGHT_GQL,
-                        variables: { input, id: oldInsight.id },
-                    })
-                )
-            }
-
-            case InsightType.CaptureGroup: {
-                const input = getCaptureGroupInsightUpdateInput(insight)
-
-                return from(
-                    this.apolloClient.mutate<UpdateLineChartSearchInsightResult>({
-                        mutation: UPDATE_LINE_CHART_SEARCH_INSIGHT_GQL,
-                        variables: { input, id: oldInsight.id },
-                    })
-                )
-            }
-
-            case InsightType.LangStats: {
-                return from(
-                    this.apolloClient.mutate<UpdateLangStatsInsightResult, UpdateLangStatsInsightVariables>({
-                        mutation: UPDATE_LANG_STATS_INSIGHT_GQL,
-                        variables: {
-                            id: oldInsight.id,
-                            input: {
-                                query: '',
-                                repositoryScope: { repositories: [insight.repository] },
-                                presentationOptions: {
-                                    title: insight.title,
-                                    otherThreshold: insight.otherThreshold,
-                                },
-                            },
-                        },
-                    })
-                )
-            }
-        }
-    }
+    public updateInsight = (input: InsightUpdateInput): Observable<unknown> => updateInsight(this.apolloClient, input)
 
     public deleteInsight = (insightId: string): Observable<unknown> =>
         from(
