+++ client/web/src/enterprise/insights/core/backend/gql-api/deserialization/create-insight-view.ts
@@ -1,72 +1,50 @@
 import { Duration } from 'date-fns'
 import { uniq } from 'lodash'
 
-import { isDefined } from '@sourcegraph/shared/src/util/types'
-
 import { InsightViewNode, TimeIntervalStepInput, TimeIntervalStepUnit } from '../../../../../../graphql-operations'
 import { Insight, InsightExecutionType, InsightType } from '../../../types'
 
-function getDurationFromStep(step: TimeIntervalStepInput): Duration {
-    switch (step.unit) {
-        case TimeIntervalStepUnit.HOUR:
-            return { hours: step.value }
-        case TimeIntervalStepUnit.DAY:
-            return { days: step.value }
-        case TimeIntervalStepUnit.WEEK:
-            return { weeks: step.value }
-        case TimeIntervalStepUnit.MONTH:
-            return { months: step.value }
-        case TimeIntervalStepUnit.YEAR:
-            return { years: step.value }
-    }
-}
-
-/**
- * Returns tuple with gql model time interval unit and value. Used to convert FE model
- * insight data series time step to GQL time interval model.
- */
-export function getStepInterval(step: Duration): [TimeIntervalStepUnit, number] {
-    const castUnits = (Object.keys(step) as (keyof Duration)[])
-        .map<[TimeIntervalStepUnit, number] | null>(key => {
-            switch (key) {
-                case 'hours':
-                    return [TimeIntervalStepUnit.HOUR, step[key] ?? 0]
-                case 'days':
-                    return [TimeIntervalStepUnit.DAY, step[key] ?? 0]
-                case 'weeks':
-                    return [TimeIntervalStepUnit.WEEK, step[key] ?? 0]
-                case 'months':
-                    return [TimeIntervalStepUnit.MONTH, step[key] ?? 0]
-                case 'years':
-                    return [TimeIntervalStepUnit.YEAR, step[key] ?? 0]
-            }
-
-            return null
-        })
-        .filter(isDefined)
-
-    if (castUnits.length === 0) {
-        throw new Error('Wrong time step format')
-    }
-
-    // Return first valid match
-    return castUnits[0]
-}
-
 /**
  * Transforms/casts gql api insight model to FE insight model. We still
- * need to have specific FE model in order to support setting-based api
+ * need to have a specific FE model in order to support setting-based api
  * approach. This transformer will be removed as soon as we sunset setting-based
  * api for insights.
- *
- * @param insight - gql insight model
  */
-export const getInsightView = (insight: InsightViewNode): Insight => {
+export const createInsightView = (insight: InsightViewNode): Insight => {
     switch (insight.presentation.__typename) {
         case 'LineChartInsightViewPresentation': {
             const isBackendInsight = insight.dataSeriesDefinitions.every(series => series.isCalculated)
-
+            const isCaptureGroupInsight = insight.dataSeriesDefinitions.some(
+                series => series.generatedFromCaptureGroups
+            )
+            // We do not support different time scope for different series at the moment
             const step = getDurationFromStep(insight.dataSeriesDefinitions[0].timeScope)
+            const repositories = uniq(
+                insight.dataSeriesDefinitions.flatMap(series => series.repositoryScope.repositories)
+            )
+
+            if (isCaptureGroupInsight) {
+                // It's safe because capture group insight always has only 1 data series
+                const { query } = insight.dataSeriesDefinitions[0] ?? {}
+
+                return {
+                    id: insight.id,
+                    type: InsightExecutionType.Backend,
+                    viewType: InsightType.CaptureGroup,
+                    title: insight.presentation.title,
+                    repositories,
+                    query,
+                    step,
+                    // In gql api we don't have this concept as visibility on FE.
+                    // Insights have special system about visibility on BE only.
+                    visibility: '',
+                    filters: {
+                        includeRepoRegexp: insight.appliedFilters.includeRepoRegex ?? '',
+                        excludeRepoRegexp: insight.appliedFilters.excludeRepoRegex ?? '',
+                    },
+                }
+            }
+
             const series = insight.presentation.seriesPresentation.map(series => ({
                 id: series.seriesId,
                 name: series.label,
@@ -99,10 +77,6 @@ export const getInsightView = (insight: InsightViewNode): Insight => {
                     },
                 }
             }
-
-            const repositories = uniq(
-                insight.dataSeriesDefinitions.flatMap(series => series.repositoryScope.repositories)
-            )
 
             return {
                 id: insight.id,
@@ -138,5 +112,20 @@ export const getInsightView = (insight: InsightViewNode): Insight => {
                 visibility: '',
             }
         }
+    }
+}
+
+function getDurationFromStep(step: TimeIntervalStepInput): Duration {
+    switch (step.unit) {
+        case TimeIntervalStepUnit.HOUR:
+            return { hours: step.value }
+        case TimeIntervalStepUnit.DAY:
+            return { days: step.value }
+        case TimeIntervalStepUnit.WEEK:
+            return { weeks: step.value }
+        case TimeIntervalStepUnit.MONTH:
+            return { months: step.value }
+        case TimeIntervalStepUnit.YEAR:
+            return { years: step.value }
     }
 }
