+++ client/web/src/enterprise/insights/core/backend/gql-api/gql/GetInsightView.ts
@@ -17,6 +17,16 @@ const INSIGHT_DATA_SERIES_FRAGMENT = gql`
     }
 `
 
+const INSIGHT_DATA_NODE_FRAGMENT = gql`
+    fragment InsightDataNode on InsightView {
+        id
+        dataSeries {
+            ...InsightDataSeries
+        }
+    }
+    ${INSIGHT_DATA_SERIES_FRAGMENT}
+`
+
 /**
  * GQL query for fetching insight data model with data series points and chart
  * information.
@@ -25,12 +35,9 @@ export const GET_INSIGHT_VIEW_GQL = gql`
     query GetInsightView($id: ID, $filters: InsightViewFiltersInput) {
         insightViews(id: $id, filters: $filters) {
             nodes {
-                id
-                dataSeries {
-                    ...InsightDataSeries
-                }
+                ...InsightDataNode
             }
         }
     }
-    ${INSIGHT_DATA_SERIES_FRAGMENT}
+    ${INSIGHT_DATA_NODE_FRAGMENT}
 `
