+++ client/web/src/enterprise/insights/core/backend/gql-api/gql/GetInsights.ts
@@ -40,6 +40,7 @@ export const INSIGHT_VIEW_FRAGMENT = gql`
                     }
                 }
                 isCalculated
+                generatedFromCaptureGroups
             }
         }
     }
